backend/requirements.txt
Django==1.8.5
djangorestframework==3.2.4
django-rest-auth==0.5.0
django-allauth==0.23.0
Markdown==2.6.2
django-filter==0.11.0
pytz==2015.6
